fastapi_server/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Email
redis>=5.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Scheduling
apscheduler>=3.10.4

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Git operations
GitPython>=3.1.40
